requirements.txt
flask==3.0.2
requests==2.31.0
python-whois==0.8.0
dnspython==2.6.1
python-nmap==0.7.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
flask-socketio==5.3.6
openai==1.12.0
python-json-logger==2.0.7
aiohttp==3.9.3
cryptography==42.0.5
pyOpenSSL==24.0.0
urllib3==2.2.1
tqdm==4.66.2
pandas==2.2.1
numpy==1.26.4 